feature(events): extended event list to support directories, save events by their id
This commit adds the ability to add events sorted by directories generated by
https://github.com/Malex14/hio_timetable_extractor. Each directory consists of a name
and potentially containing events and/or subdirectories. Events are stored in the
userconfig by their id. The event's name is now being saved in the corresponding
eventDetails. The filter functionality has been updated to support directories as well.